requirements.txt
beautifulsoup4
requests
python-dotenv
pandas


sentence-transformers
scikit-learn


google-api-python-client
google-auth-oauthlib
google-auth-httplib2